requirements.txt
RPi.GPIO
picamera
#git+https://github.com/adafruit/Adafruit_Python_DHT

